chore(cloud):  remove old nav entirely
