+++ ui/apps/dashboard/src/components/Navigation/Announcements/announcements.ts
