+++ ui/apps/dashboard/src/components/Navigation/KeysNavItem.tsx
