+++ ui/apps/dashboard/src/components/Navigation/navItems.ts
