--- ui/apps/dashboard/src/components/NavigationV2/Environments.tsx
@@ -1,294 +1,0 @@
-import { useEffect, useState } from 'react';
-
-import { Listbox } from '@headlessui/react';
-import { OptionalTooltip } from '@inngest/components/Tooltip/OptionalTooltip';
-import {
-  Tooltip,
-  TooltipContent,
-  TooltipTrigger,
-} from '@inngest/components/Tooltip/Tooltip';
-import { cn } from '@inngest/components/utils/classNames';
-import {
-  RiCloudFill,
-  RiCloudLine,
-  RiErrorWarningLine,
-  RiExpandUpDownLine,
-  RiLoopLeftLine,
-} from '@remixicon/react';
-import { useLocation, useNavigate } from '@tanstack/react-router';
-
-import { useEnvironments } from '@/queries';
-import {
-  EnvironmentType,
-  getDefaultEnvironment,
-  getSortedBranchEnvironments,
-  getTestEnvironments,
-  type Environment,
-} from '@/utils/environments';
-
-// Some URLs cannot just swap between environments,
-// we need to redirect to a less specific resource URL that is shared across environments
-// for the user to switch context correctly
-const useSwitchablePathname = (): string => {
-  const location = useLocation();
-  const pathname = location.pathname;
-
-  //
-  // Parse pathname into segments and strip env/envSlug prefix
-  // to match Next.js useSelectedLayoutSegments behavior
-  let segments = pathname.split('/').filter((segment) => segment.length > 0);
-
-  if (segments[0] === 'env') {
-    segments = segments.length > 1 ? segments.slice(2) : [];
-  }
-
-  // Accounts are not environment specific
-  if (pathname.match(/^\/settings\//)) {
-    return '/functions';
-  }
-
-  if (pathname.match(/^\/billing(?:\/|$)/)) {
-    return '/';
-  }
-
-  // Deploys should always move to the root resource level
-  if (segments[0] === 'apps') {
-    return '/apps';
-  }
-  // Manage paths, we drop the id at the end
-  if (segments[0] === 'manage') {
-    return '/' + segments.slice(0, 2).join('/');
-  }
-
-  // Logs are specific to a given environment, return to the function dashboard
-  if (segments[0] === 'functions' && segments[2] === 'logs') {
-    return '/' + segments.slice(0, 3).join('/');
-  }
-
-  if (segments.length === 0) {
-    return '/';
-  }
-
-  return '/' + segments.join('/');
-};
-
-const selectedName = (name: string, collapsed: boolean) => {
-  switch (name) {
-    case 'Production':
-      return collapsed ? 'PR' : name;
-    case 'Branch Environments':
-      return collapsed ? 'BE' : name;
-    default:
-      return collapsed ? name.substring(0, 2).toUpperCase() : name;
-  }
-};
-
-const SelectedDisplay = ({
-  selected,
-  collapsed,
-}: {
-  selected: Environment | null;
-  collapsed: boolean;
-}) => (
-  <span
-    className={`flex h-full flex-row items-center leading-none ${
-      collapsed ? '' : 'min-w-0 truncate'
-    }`}
-  >
-    {selected ? (
-      <span className="flex h-full items-center leading-none">
-        {selected.type === EnvironmentType.BranchParent
-          ? selectedName('Branch Environments', collapsed)
-          : selectedName(selected.name, collapsed)}
-      </span>
-    ) : (
-      <>
-        {!collapsed && <RiCloudLine className="mr-2 h-4 w-4" />}
-        <span className="flex h-full items-center leading-none">
-          {selectedName('All Environments', collapsed)}
-        </span>
-      </>
-    )}
-  </span>
-);
-
-const tooltip = (selected: Environment | null) =>
-  !selected
-    ? 'All Environments'
-    : selected.type === EnvironmentType.BranchParent
-    ? 'Branch Environments'
-    : selected.name;
-
-type EnvironmentSelectMenuProps = {
-  activeEnv?: Environment;
-  collapsed: boolean;
-};
-
-export default function EnvironmentSelectMenu({
-  activeEnv,
-  collapsed,
-}: EnvironmentSelectMenuProps) {
-  const navigate = useNavigate();
-  type ListboxValue = Environment | null | 'view_all' | 'sync_branch';
-  const [selected, setSelected] = useState<ListboxValue>(null);
-  const nextPathname = useSwitchablePathname();
-  const [{ data: envs = [], error }] = useEnvironments();
-
-  //
-  // Sync selected state with activeEnv from route
-  useEffect(() => {
-    setSelected(activeEnv || null);
-  }, [activeEnv]);
-
-  if (error) {
-    console.error('error fetching envs', error);
-    return (
-      <Tooltip>
-        <TooltipTrigger asChild>
-          <div className="bg-error text-error flex h-8 w-full items-center justify-start gap-x-2 rounded px-2">
-            <RiErrorWarningLine className="w-4" />
-            {!collapsed && <div>Env Error</div>}
-          </div>
-        </TooltipTrigger>
-        <TooltipContent
-          side="right"
-          className="text-error bg-error rounded text-xs"
-        >
-          Error loading environments. Please try again or contact support if the
-          issue does not resolve.
-        </TooltipContent>
-      </Tooltip>
-    );
-  }
-
-  const defaultEnvironment = getDefaultEnvironment(envs);
-  const includeArchived = false;
-  const mostRecentlyCreatedBranchEnvironments = getSortedBranchEnvironments(
-    envs,
-    includeArchived,
-  ).slice(0, 5);
-  const testEnvironments = getTestEnvironments(envs, includeArchived);
-
-  const selectedEnv = typeof selected === 'string' ? null : selected;
-
-  const onSelect = (value: ListboxValue) => {
-    if (value === null || value === 'view_all' || value === 'sync_branch') {
-      navigate({ to: '/env' });
-      return;
-    }
-
-    navigate({
-      to: `/env/${value.slug}${nextPathname}`,
-    });
-  };
-
-  return (
-    <Listbox value={selected} onChange={onSelect}>
-      {({ open }) => (
-        <div className="relative flex h-8 items-center">
-          <OptionalTooltip tooltip={collapsed && tooltip(selectedEnv)}>
-            {/* Trigger styled inline to match the breadcrumb look in the
-                top bar: flat, transparent, hover:bg-canvasMuted, text-basis.
-                If the env switcher ever gets mounted somewhere else we'll
-                want to introduce a variant prop here. */}
-            <Listbox.Button
-              className={`text-basis hover:bg-canvasMuted flex h-8 items-center gap-1.5 rounded px-2 text-sm leading-none ${
-                collapsed ? 'w-8 justify-center' : ''
-              } ${open ? 'bg-canvasMuted' : ''}`}
-            >
-              <SelectedDisplay selected={selectedEnv} collapsed={collapsed} />
-              {!collapsed && (
-                <RiExpandUpDownLine
-                  className="text-muted h-4 w-4 shrink-0"
-                  aria-hidden="true"
-                />
-              )}
-            </Listbox.Button>
-          </OptionalTooltip>
-
-          <Listbox.Options className="bg-canvasBase border-subtle absolute top-10 z-50 max-h-[calc(100vh-8rem)] w-[250px] divide-none overflow-y-auto rounded border shadow focus:outline-none">
-            {defaultEnvironment !== null && (
-              <EnvironmentItem environment={defaultEnvironment} />
-            )}
-
-            {testEnvironments.length > 0 &&
-              testEnvironments.map((env) => (
-                <EnvironmentItem key={env.id} environment={env} />
-              ))}
-
-            <div>
-              <div className="bg-canvasBase text-disabled border-subtle flex h-[18px] cursor-not-allowed items-center gap-3 border-t px-3 py-4 text-xs font-normal">
-                Branch Environments
-              </div>
-              {mostRecentlyCreatedBranchEnvironments.length > 0 ? (
-                mostRecentlyCreatedBranchEnvironments.map((env) => (
-                  <EnvironmentItem
-                    key={env.id}
-                    environment={env}
-                    variant="compact"
-                  />
-                ))
-              ) : (
-                <Listbox.Option
-                  key="sync-branch"
-                  value="sync_branch"
-                  className="bg-canvasBase hover:bg-canvasSubtle text-subtle flex h-10 cursor-pointer items-center gap-3 px-3 text-[13px] font-normal"
-                >
-                  <RiLoopLeftLine className="h-3 w-3" />
-                  Sync a branch
-                </Listbox.Option>
-              )}
-            </div>
-
-            <div>
-              <Listbox.Option
-                key="view-all"
-                value="view_all"
-                className="hover:bg-canvasSubtle text-subtle flex h-10 cursor-pointer items-center gap-3 whitespace-nowrap px-3 text-[13px] font-normal"
-              >
-                <RiCloudFill className="h-3 w-3" />
-                View All Environments
-              </Listbox.Option>
-            </div>
-          </Listbox.Options>
-        </div>
-      )}
-    </Listbox>
-  );
-}
-
-function EnvironmentItem({
-  environment,
-  name,
-  variant = 'normal',
-}: {
-  environment: Environment;
-  name?: string;
-  variant?: 'compact' | 'normal';
-}) {
-  let statusColorClass: string;
-  if (environment.isArchived) {
-    statusColorClass = 'bg-surfaceMuted';
-  } else {
-    statusColorClass = 'bg-primary-moderate';
-  }
-
-  return (
-    <Listbox.Option
-      key={environment.id}
-      value={environment}
-      className={cn(
-        'bg-canvasBase hover:bg-canvasSubtle text-subtle flex h-10 cursor-pointer items-center gap-3 px-3 text-[13px] font-normal',
-        variant === 'compact' && 'py-2',
-      )}
-    >
-      <span
-        className={cn(
-          'block h-1.5 w-1.5 shrink-0 rounded-full',
-          statusColorClass,
-        )}
-      />
-      <span className="truncate">{name || environment.name}</span>
-    </Listbox.Option>
-  );
-}
