--- ui/apps/dashboard/src/components/NavigationV2/QuickSearch/hooks.ts
@@ -1,17 +1,0 @@
-import { useEffect, useState } from 'react';
-
-export function useDebounce<T>(value: T, delay: number) {
-  const [debouncedValue, setDebouncedValue] = useState(value);
-
-  useEffect(() => {
-    const handler = setTimeout(() => {
-      setDebouncedValue(value);
-    }, delay);
-
-    return () => {
-      clearTimeout(handler);
-    };
-  }, [value, delay]);
-
-  return debouncedValue;
-}
