+++ ui/apps/dashboard/src/routes/_authed.tsx
@@ -7,9 +7,7 @@ import {
   useMatch,
 } from '@tanstack/react-router';
 
-import LayoutV1 from '@/components/Layout/Layout';
-import LayoutV2 from '@/components/Layout/LayoutV2';
-import { useNavigationV2 } from '@/components/Layout/useNavigationV2';
+import Layout from '@/components/Layout/Layout';
 import { navCollapsed } from '@/lib/nav';
 import { getProfileDisplay } from '@/queries/server/profile';
 import NotFound from '@/components/Error/NotFound';
@@ -56,7 +54,6 @@ export const Route = createFileRoute('/_authed')({
 
 function Authed() {
   const { navCollapsed, profile } = Route.useLoaderData();
-  const Layout = useNavigationV2() ? LayoutV2 : LayoutV1;
   const activeEnv = useMatch({
     from: '/_authed/env/$envSlug',
     shouldThrow: false,
